src/requirements.txt
requests==2.24.0
requests-oauthlib==1.3.0
pyyaml==5.3.1